[ADD]Added code for General feedback : 169, 171, 172, 173, 174 and 178 and Investment feedback : 276, 312, 319, 321, 322, 327, 329, 330

+++ jt_budget_mgmt/models/control_assigned_amounts.py
@@ -818,8 +818,8 @@ class ControlAssignedAmountsLines(models.Model):
     end_date = fields.Date(string='End date')
 
     # authorized = fields.Integer(string='Authorized')
-    assigned = fields.Integer(string='Assigned')
-    available = fields.Integer(string='Available')
+    assigned = fields.Float(string='Assigned')
+    available = fields.Float(string='Available')
     currency_id = fields.Many2one(
         'res.currency', default=lambda self: self.env.company.currency_id)
